requirements.txt
docker==7.1.0
unittest-xml-reporting==3.2.0
fake-bpy-module-latest==20231010
PyGithub==2.1.1
debugpy==1.8.1
numpy==1.26.4
websocket==0.2.1
python-dotenv==1.0.1

# documenation 
mkdocs<=1.5.3
mkdocs-material<=9.5.13
mkdocs-material-extensions<=1.3.1
mkdocs-mermaid2-plugin<=1.1.1
mkdocs-monorepo-plugin<=1.1.0
termynal<=0.12.1
mkdocstrings[python]<=1.9.0